app/requirements.txt
pymongo==4.10.1
passlib==1.7.4
bcrypt==4.0.1
fastapi[standard]==0.115.4
python-jose==3.3.0
python-dotenv==1.0.1